Add simple tests for GET /
A fundamental operation for most OpenStack services is providing
information about what versions of an API are available to clients.
A version document can be retrieved by sending an unauthenticated
GET request to the root URL ("/") for most services, including
Neutron.  This capability is important for discovery in that clients
can learn how to interact with the cloud in question, and DefCore
considers it an important capability for interoperability and has
added similar capabilities to it's Guidelines for other services.[1][2]
As Neutron moves toward microversioning [3], being able to retrieve
a version document will be increasingly important for clients.
However, there are currently no tests for GET /, so DefCore cannot
make this a required Capability.

This patch adds two simple tests for GET / that perform simple
checks on document structure and version information.

[1] http://git.openstack.org/cgit/openstack/defcore/tree/2016.01.json#n117
[2] http://git.openstack.org/cgit/openstack/defcore/tree/2016.01.json#n1379
[3] https://etherpad.openstack.org/p/newton-neutron-future-neutron-api